sshfs-fuse: bump 2.8 -> 2.10
Dropped the Makefile.am patch we were carrying because it was merged
upstream in the 2.10 release.

sshfs 2.x are maintenance releases which will not see significant
developments/features added. They are mantained for libfuse v2
compatibility for those who can't upgrade to fuse3/sshfs3. The last
and only big feature of the 2.x branch is migration to meson (under
the new maintainership all libfuse projects moved to meson).

Support for building sshfs with autotools was dropped in sshfs v3
so we also build v2.10 with meson because it's actively maintained.

Signed-off-by: Ioan-Adrian Ratiu <adi@adirat.com>