sshfs-fuse: bump 2.8 -> 2.10
sshfs 2.x are maintenance releases which will not see significant
development or features added. They are just mantained for libfuse 2
compatibility for those who can't upgrade to fuse3 / sshfs 3. The last
and only big feature of the 2.x branch is migration to meson (under
the new maintainership all libfuse projects move to meson apparently).

Support for building sshfs with autotools was dropped in sshfs 3
so we also build v2.10 with meson here because it's more actively
maintained.

Also dropped the Makefile.am patch we were carrying because:
  1. It was integrated upstream in the 2.x branch
  2. We don't use the logic anymore since building with meson.

Signed-off-by: Ioan-Adrian Ratiu <adi@adirat.com>