fuse: add fuse 3.2.2 beside 2.9.7 maintenance release
libfuse v3 is a backwards incompatible release (new generation of
maintainers) meant to be installed side by side with v2 which is
deprecated but still expected to live on for a few years until all
dependent software gets ported to v3.

Quoting the upstream note to packagers [1]:

"libfuse 3 is designed to be co-installable with libfuse 2. However,
some files will be installed by both libfuse 2 and libfuse
3 (e.g. /etc/fuse.conf, the udev and init scripts, and the
mount.fuse(8) manpage). These files should be taken from libfuse
3. The format/content is guaranteed to remain backwards compatible
with libfuse 2.

We recommend to ship libfuse2 and libfuse3 in three separate packages:
a libfuse-common package that contains files shared by libfuse
2+3 (taken from the libfuse3 tarball), and libfuse2 and libfuse3
packages that contain the shared library and helper programs for the
respective version."

[1] https://github.com/libfuse/libfuse/releases/tag/fuse-3.0.0

Signed-off-by: Ioan-Adrian Ratiu <adi@adirat.com>