sshfs-fuse: add sshfs v3.3.1
This adds the latest release of sshfs but v2.10 is also kept around
because some users might be stuck on libfuse 2.

Considering that libfuse 3 can be installed alongside 2 and upstream
recommends everyone to upgrade to sshfs 3, we can drop v2.10 when this
newer version has sufficient testing and there are no additional
reasons to keep the old one around (basically what I'm trying to do is
avoid forcing sshfs users to install libfuse 3).

Signed-off-by: Ioan-Adrian Ratiu <adi@adirat.com>